`senator-minor-mode-setup' now uses the `semantic-active-p' function to check if the current buffer is set up for parsing.
`senator-full-token-name' now append the parent names in reverse order
to the token name.  Thus the `senator-jump' completion list is now
easy to use.  You can just type a partial token name and press ENTER
to jump to it.  If the name is not unique a completion buffer is
displayed allowing to choose the right one.

`senator-beginning-of-defun' and `senator-end-of-defun' now
respectively step at start or end of tokens with identifier specified
in `senator-step-at-token-ids'.  Thus, adviced `beginning-of-defun'
and `end-of-defun' now work with any tokens where you can step at.

`senator-step-at-start-end-token-ids' can be set to `t' to step at
start and end of any token where it is allowed to step.