(senator-minor-mode-name): Can be set to nil to disable display of the minor mode name in the mode line.
(senator-minor-mode-isearch-suffix): New option to set up the string
appended to the mode name when senator isearch mode is on.

(senator-minor-mode-hook, senator-minor-mode-on-hook,
senator-minor-mode-off-hook): Moved to group minor mode options
together.

(senator-menu-bar): For consistency always use "Senator" for the menu
title.

(senator-show-status): Use `senator-minor-mode-name' and
`senator-minor-mode-isearch-suffix' settings.

(senator-minor-mode-setup): Signal an error when attempting to turn on
Senator minor mode in a buffer not set up for parsing.

(senator-show-status, senator-mode-line-toggle): New GNU Emacs 21
goodies to turn off Senator minor mode from the mode line or to turn
it on or off from the the minor mode menu in the mode line.